requirements.txt
langflow~=1.1.0
python-dotenv~=1.0.1
requests~=2.32.3
streamlit~=1.40.1
astrapy~=1.5.2